+++ rofr/src/header.rs
@@ -8,9 +8,3 @@ pub const SERVICE_UID: &str = "RoFr-Service-Uid";
 pub const REQUEST_ID: &str = "RoFr-Request-Id";
 /// Endpoint request id header.
 pub const MESSAGE_ID: &str = "RoFr-Message-Id";
-/// Endpoint request schema header.
-pub const REQUEST_SCHEMA: &str = "RoFr-Request-Schema";
-/// Endpoint response schema header.
-pub const RESPONSE_SCHEMA: &str = "RoFr-Response-Schema";
-/// Stream message schema header.
-pub const MESSAGE_SCHEMA: &str = "RoFr-Message-Schema";