+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Remove the need for `serde_json` as a dependency in implementations.
 - Remove the need for `ulid` as a dependency in implementations.
+- `schemars` and schema headers removed.
 
 ## v0.1.0